backend/requirements.txt
# -----------------------------------------------
# Fichier des dépendances du projet e-mpj
# -----------------------------------------------
# Mis à jour le: 26/03/2024
# 
# Ce fichier a été généré automatiquement à partir de
# l'environnement de développement actif (pip freeze)
# et complété avec les bibliothèques détectées dans le code
# -----------------------------------------------

# Django et composants principaux
Django==5.1.7
asgiref==3.8.1
sqlparse==0.5.3
tzdata==2025.1

# API REST
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
drf-yasg==1.21.7

# Gestion des variables d'environnement
django-environ==0.12.0
python-dotenv==1.0.1
python-decouple==3.8

# Serveur de production
gunicorn==23.0.0

# Base de données
psycopg2-binary==2.9.10
dj-database-url==2.3.0
django-filter==23.5

# Gestion des fichiers et images
Pillow==11.1.0
django-storages==1.14.5
cloudinary==1.38.0
django-cloudinary-storage==0.3.0
boto3==1.34.34

# Authentification et sécurité
django-allauth==65.5.0
django-cors-headers==4.7.0

# Optimisation et performance
whitenoise==6.9.0
redis==5.2.1
django-redis==5.4.0

# Tâches asynchrones
celery==5.4.0
amqp==5.3.1
billiard==4.2.1
kombu==5.5.1
vine==5.1.0
flower==2.0.1
prometheus_client==0.21.1
tornado==6.4.2
humanize==4.12.2
apscheduler==3.10.4

# Frontend et UI
django-tailwind==3.8.0
django-browser-reload==1.18.0
django-crispy-forms==2.1

# Internationalisation
Babel==2.17.0
pytz==2025.2
django-modeltranslation==0.19.13

# Outils de développement
django-debug-toolbar==5.1.0
django-extensions==3.2.3
typing_extensions==4.12.2

# HTTP et réseau
requests==2.31.0
chardet==5.2.0
idna==3.7
urllib3==2.2.2
certifi==2024.6.2

# CLI et utilitaires
click==8.1.8
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
prompt_toolkit==3.0.50
wcwidth==0.2.13
python-dateutil==2.9.0.post0
six==1.17.0

# Tests
pytest==8.3.5
pytest-django==4.10.0
coverage==7.7.1
iniconfig==2.1.0
packaging==24.2
pluggy==1.5.0

# Google Sheets et authentification
gspread==6.0.2
google-auth==2.28.1

# Génération de rapports et manipulation de fichiers
reportlab==4.1.0
openpyxl==3.1.2


